SqlQuoter moved from Sql component folder to Doctrine component folder
- class is related to entity manager which is Doctrine components responsibility
